requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
cryptography>=41.0.0
pandas>=2.0.0
pillow>=10.0.0
